+++ packages/ui/src/index.ts
@@ -49,3 +49,6 @@ export * from './Box';
 export * from './Container';
 export * from './Grid';
 export * from './Stack';
+
+export * from './DataGrid';
+export * from './DatePicker';